+++ src/simulation/data_structure.jl
@@ -21,7 +21,7 @@ mutable struct Station
     max_charging_rate_per_charging_point_fast
 end
 
-Station(station_node_id::Integer, initial_cars_number::Integer, initial_id::Integer) = begin
+Station(station_node_id::Int64, initial_cars_number::Int64, initial_id::Int64) = begin
     #get teh different information from the graph
     prop_dict = props(manhaten_city_graph, station_node_id)
     
@@ -36,15 +36,15 @@ Station(station_node_id::Integer, initial_cars_number::Integer, initial_id::Inte
     id = initial_id
     
     cars = DataFrame(car_id = collect(id:id+initial_cars_number-1), car_type = [Smart_ED for i in 1:initial_cars_number],
-                    status = CAR_PARKED .* ones(Integer, initial_cars_number),
+                    status = CAR_PARKED .* ones(Int64, initial_cars_number),
                     last_battery_level = 100.0 .* ones(Float64, initial_cars_number), start_charging_time = NaN * ones(initial_cars_number),
                     start_reservation_time = NaN * ones(initial_cars_number), pending_reservation = zeros(initial_cars_number), expected_arrival_time = NaN * ones(initial_cars_number))
     
     #create the parking places
     parking_places = DataFrame(p_id = collect(1:total_parking_places), 
-                                status = vcat(P_OCCUPIED .* ones(Integer, initial_cars_number), P_FREE .* ones(Integer, total_parking_places - initial_cars_number)), 
-                                cars = vcat(collect(id:id+initial_cars_number-1), -1 .* ones(Integer, total_parking_places - initial_cars_number)),
-                                pending_reservation = zeros(Integer, total_parking_places) )
+                                status = vcat(P_OCCUPIED .* ones(Int64, initial_cars_number), P_FREE .* ones(Int64, total_parking_places - initial_cars_number)), 
+                                cars = vcat(collect(id:id+initial_cars_number-1), -1 .* ones(Int64, total_parking_places - initial_cars_number)),
+                                pending_reservation = zeros(Int64, total_parking_places) )
     
     
     Station(cars, parking_places, total_parking_places, max_power, charging_station_base_cost, charging_point_cost_slow,
@@ -54,10 +54,11 @@ end
 mutable struct Solution
     # decision variables
     open_stations_state::Array{Bool, 1}
-    initial_cars_number::Array{Integer, 1}
+    initial_cars_number::Array{Int64, 1}
     selected_paths::Array{Array{Bool,1}, 1}
 end
 Solution() = Solution(falses(length(get_potential_locations())), zeros(length(get_potential_locations())), []) 
+
 
 
 mutable struct Scenario
